+++ integration_tests/piggy_bank/scripts/build_wasm.sh
@@ -8,5 +8,5 @@ for smart_contract_json in $SMART_CONTRACT_JSONS
 do
     smart_contract_folder=$(dirname $smart_contract_json)
     echo ""
-    (set -x; mxpy --verbose contract build $smart_contract_folder)
+    (set -x; mxpy contract build --path $smart_contract_folder)
 done