+++ tests/conftest.py
@@ -33,6 +33,7 @@ def scenario_data():
         InternalContractData(
             contract_id=contract_id,
             address=address,
+            serializer=None,
             wasm_hash=wasm_hash,
             deploy_time=1,
             last_upgrade_time=1,